+++ src/components/shared/RentInformationTable/RentInformationTable.spec.tsx
@@ -22,7 +22,7 @@ describe('RentInformationTable', () => {
     .addRentInformation(
       new RentInformationBuilder()
         .withDueDate(new MonthYear(3, 2023))
-        .withPayment(1, new Date(2023, 5, 11))
+        .withPayment(1, new Date(2023, 5, 11), 'note')
         .withBankTransferDocumentId('banktransfer2')
         .build(),
     )